business-aware-rag/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sentence-transformers==2.6.1
faiss-cpu==1.8.0
numpy>=1.24.0

openai>=1.12.0
pydantic>=2.6.0
pyyaml>=6.0

scikit-learn>=1.3.0